+++ piecrust/src/store.rs
@@ -10,6 +10,7 @@ mod bytecode;
 mod diff;
 mod memory;
 mod mmap;
+mod objectcode;
 mod session;
 
 use std::collections::btree_map::Entry::*;
@@ -25,6 +26,7 @@ use flate2::Compression;
 pub use bytecode::Bytecode;
 use diff::diff;
 pub use memory::Memory;
+pub use objectcode::Objectcode;
 use piecrust_uplink::ModuleId;
 pub use session::ModuleSession;
 
@@ -34,8 +36,7 @@ const BYTECODE_DIR: &str = "bytecode";
 const MEMORY_DIR: &str = "memory";
 const DIFF_EXTENSION: &str = "diff";
 const INDEX_FILE: &str = "index";
-pub const COMPILED_DIR: &str = "compiled";
-const SKIP_LIST: [&str; 1] = [COMPILED_DIR];
+const OBJECTCODE_EXTENSION: &str = ".a";
 
 type Root = [u8; ROOT_LEN];
 
@@ -65,8 +66,12 @@ impl ModuleStore {
         let commits = read_all_commits(root_dir)?;
 
         let loop_root_dir = root_dir.to_path_buf();
-        let sync_loop =
-            thread::spawn(|| sync_loop(loop_root_dir, commits, calls));
+
+        // The thread is given a name to allow for easily identifying it while
+        // debugging.
+        let sync_loop = thread::Builder::new()
+            .name(String::from("PiecrustSync"))
+            .spawn(|| sync_loop(loop_root_dir, commits, calls))?;
 
         Ok(Self {
             sync_loop,
@@ -283,7 +288,7 @@ pub(crate) struct Commit {
 
 pub(crate) enum Call {
     Commit {
-        modules: BTreeMap<ModuleId, (Bytecode, Memory)>,
+        modules: BTreeMap<ModuleId, (Bytecode, Objectcode, Memory)>,
         base: Option<(Root, Commit)>,
         replier: mpsc::SyncSender<io::Result<(Root, Commit)>>,
     },
@@ -464,7 +469,7 @@ fn write_commit<P: AsRef<Path>>(
     root_dir: P,
     commits: &mut BTreeMap<Root, Commit>,
     base: Option<(Root, Commit)>,
-    commit_modules: BTreeMap<ModuleId, (Bytecode, Memory)>,
+    commit_modules: BTreeMap<ModuleId, (Bytecode, Objectcode, Memory)>,
 ) -> io::Result<(Root, Commit)> {
     let root_dir = root_dir.as_ref();
 
@@ -503,7 +508,7 @@ fn write_commit_inner<P: AsRef<Path>>(
     commit_dir: P,
     base: Option<(Root, Commit)>,
     modules: BTreeMap<ModuleId, Root>,
-    commit_modules: BTreeMap<ModuleId, (Bytecode, Memory)>,
+    commit_modules: BTreeMap<ModuleId, (Bytecode, Objectcode, Memory)>,
 ) -> io::Result<Commit> {
     let root_dir = root_dir.as_ref();
     let commit_dir = commit_dir.as_ref();
@@ -518,13 +523,16 @@ fn write_commit_inner<P: AsRef<Path>>(
 
     match base {
         None => {
-            for (module, (bytecode, memory)) in commit_modules {
+            for (module, (bytecode, objectcode, memory)) in commit_modules {
                 let module_hex = hex::encode(module);
 
                 let bytecode_path = bytecode_dir.join(&module_hex);
+                let objectcode_path =
+                    bytecode_path.with_extension(OBJECTCODE_EXTENSION);
                 let memory_path = memory_dir.join(&module_hex);
 
                 fs::write(bytecode_path, &bytecode)?;
+                fs::write(objectcode_path, &objectcode)?;
                 fs::write(memory_path, &memory.read())?;
             }
         }
@@ -539,12 +547,17 @@ fn write_commit_inner<P: AsRef<Path>>(
                 let module_hex = hex::encode(module);
 
                 let bytecode_path = bytecode_dir.join(&module_hex);
+                let objectcode_path =
+                    bytecode_path.with_extension(OBJECTCODE_EXTENSION);
                 let memory_path = memory_dir.join(&module_hex);
 
                 let base_bytecode_path = base_bytecode_dir.join(&module_hex);
+                let base_objectcode_path =
+                    base_bytecode_path.with_extension(OBJECTCODE_EXTENSION);
                 let base_memory_path = base_memory_dir.join(&module_hex);
 
                 fs::hard_link(base_bytecode_path, bytecode_path)?;
+                fs::hard_link(base_objectcode_path, objectcode_path)?;
                 fs::hard_link(&base_memory_path, &memory_path)?;
 
                 // If there is a diff of this memory in the base module, and it
@@ -561,7 +574,7 @@ fn write_commit_inner<P: AsRef<Path>>(
                 }
             }
 
-            for (module, (bytecode, memory)) in commit_modules {
+            for (module, (bytecode, objectcode, memory)) in commit_modules {
                 let module_hex = hex::encode(module);
 
                 match base_commit.modules.contains_key(&module) {
@@ -590,9 +603,12 @@ fn write_commit_inner<P: AsRef<Path>>(
                     }
                     false => {
                         let bytecode_path = bytecode_dir.join(&module_hex);
+                        let objectcode_path =
+                            bytecode_path.with_extension(OBJECTCODE_EXTENSION);
                         let memory_path = memory_dir.join(&module_hex);
 
                         fs::write(bytecode_path, &bytecode)?;
+                        fs::write(objectcode_path, &objectcode)?;
                         fs::write(memory_path, memory.read())?;
                     }
                 }
@@ -666,14 +682,14 @@ fn compute_root<'a, I>(
     modules: I,
 ) -> (Root, BTreeMap<ModuleId, Root>)
 where
-    I: IntoIterator<Item = (&'a ModuleId, &'a (Bytecode, Memory))>,
+    I: IntoIterator<Item = (&'a ModuleId, &'a (Bytecode, Objectcode, Memory))>,
 {
     let iter = modules.into_iter();
 
     let mut leaves_map = BTreeMap::new();
 
     // Compute the hashes of changed memories
-    for (module, (_, memory)) in iter {
+    for (module, (_, _, memory)) in iter {
         let mut hasher = blake3::Hasher::new();
         hasher.update(&memory.read());
         leaves_map.insert(*module, Root::from(hasher.finalize()));